- Added Replace to LinkedList ADT - Added LinkedListInterface - Finished Donee Subsystem

+++ Assignment_test/src/DoneeSubsystem/Test.java
@@ -35,80 +35,80 @@ public class Test {
                 System.out.println("4. Search donee details");
                 System.out.println("5. List donees");
                 System.out.println("6. Filter donee based on category");
-//                System.out.println("7. Generate summary reports");
-                System.out.println("8. Save Changes");
-                System.out.println("9. Exit");
+                System.out.println("7. Save Changes");
+                System.out.println("8. Exit");
                 System.out.print("\nEnter your choice: ");
 
-                int choice = scanner.nextInt();
-                scanner.nextLine(); // consume the newline
+                String choice = scanner.nextLine();
                 System.out.println("");
 
                 switch (choice) {
-                    case 1 -> {
+                    case "1" -> {
                         // Add a new donee
-                        System.out.println("Enter donee type: ");
-                        System.out.println("1. Individual ");
-                        System.out.println("2. Family ");
-                        System.out.println("3. Organization \n");
-                        System.out.print("Enter your choice: ");
-                        int type = scanner.nextInt();
-                        scanner.nextLine(); // consume the newline
-
                         do {
-                            switch (type) {
-                                case 1 -> {
+                            System.out.println("Enter donee type: ");
+                            System.out.println("1. Individual ");
+                            System.out.println("2. Family ");
+                            System.out.println("3. Organization \n");
+                            System.out.print("Enter your choice: ");
+                            choice = scanner.nextLine();
+
+                            switch (choice) {
+                                case "1" -> {
                                     String id = generateDoneeId(doneeList);
                                     System.out.print("Enter name: ");
                                     String name = scanner.nextLine();
                                     doneeList.insert(new Individual(id, name));
-                                    System.out.println(Green + "\nDonee added with ID: " + Reset + id);
+                                    System.out.println(Green + "Individual added with Donee ID: " + id + Reset);
                                 }
-                                case 2 -> {
+                                case "2" -> {
                                     String id = generateDoneeId(doneeList);
                                     System.out.print("Enter family name: ");
                                     String orgName = scanner.nextLine();
                                     doneeList.insert(new Family(id, orgName));
-                                    System.out.println(Green + "\nDonee added with ID: " + Reset + id);
+                                    System.out.println(Green + "Family added with Donee ID: " + id + Reset);
                                 }
-                                case 3 -> {
+                                case "3" -> {
                                     String id = generateDoneeId(doneeList);
                                     System.out.print("Enter organization name: ");
                                     String orgName = scanner.nextLine();
                                     doneeList.insert(new Organization(id, orgName));
-                                    System.out.println(Green + "\nDonee added with ID: " + Reset + id);
+                                    System.out.println(Green + "Organization added with Donee ID: " + id + Reset);
                                 }
                                 default -> {
                                     System.out.println(Red + "Invalid choice. Please only enter '1', '2' or '3'.\n" + Reset);
                                 }
                             }
-                        } while (type < 1 || type > 3);
+                        } while (!"1".equals(choice) && !"2".equals(choice) && !"3".equals(choice));
                     }
-
-                    case 2 -> {
+                    case "2" -> {
                         // Remove a donee
                         System.out.print("Enter donee ID to remove: ");
                         String id = scanner.nextLine();
-                        doneeList.deleteById(id);
-                        System.out.println("Donee removed.");
+                        Donee doneeToRemove = doneeList.findById(id);
+                        if (doneeToRemove != null) {
+                            doneeList.deleteById(id);
+                            System.out.println(Green + "Donee (" + doneeToRemove.getName() + ") was removed successfully." + Reset);
+                        } else {
+                            System.out.println(Red + "Invalid ID. No donee found with the given ID." + Reset);
+                        }
                     }
-                    case 3 -> {
+                    case "3" -> {
                         // Update donee details
-                        System.out.println("Enter donee ID to update: ");
+                        System.out.print("Enter donee ID to update: ");
                         String id = scanner.nextLine();
                         Donee doneeToUpdate = doneeList.findById(id);
                         if (doneeToUpdate != null) {
-                            System.out.println("\nFound donee: \n\n" + doneeToUpdate.toString());
+                            System.out.println(Green + "\nFound donee: " + Reset + doneeToUpdate.toString() + "\n");
                             System.out.println("What would you like to update?");
                             System.out.println("1. Name");
                             System.out.println("2. Category \n");
                             System.out.print("Enter your choice: ");
 
-                            int updateChoice = scanner.nextInt();
-                            scanner.nextLine();
+                            choice = scanner.nextLine();
 
-                            switch (updateChoice) {
-                                case 1 -> {
+                            switch (choice) {
+                                case "1" -> {
                                     // Update Name
                                     do {
                                         System.out.print("\nEnter new donee name: ");
@@ -118,28 +118,26 @@ public class Test {
                                         break;
                                     } while (true);
                                 }
-                                case 2 -> {
+                                case "2" -> {
                                     // Update Category
-                                    int newCategory = 0;
                                     do {
                                         System.out.println("Choose a new donee category: ");
                                         System.out.println("1. Individual ");
                                         System.out.println("2. Family ");
                                         System.out.println("3. Organization \n");
                                         System.out.print("Enter your choice: ");
-                                        newCategory = scanner.nextInt();
-                                        scanner.nextLine();
+                                        choice = scanner.nextLine();
 
-                                        switch (newCategory) {
-                                            case 1 -> {
+                                        switch (choice) {
+                                            case "1" -> {
                                                 doneeList.replace(doneeToUpdate, doneeList.changeToIndividual(doneeToUpdate));
                                                 System.out.println(Green + "Category updated to individual successfully." + Reset);
                                             }
-                                            case 2 -> {
+                                            case "2" -> {
                                                 doneeList.replace(doneeToUpdate, doneeList.changeToFamily(doneeToUpdate));
                                                 System.out.println(Green + "Category updated to family successfully." + Reset);
                                             }
-                                            case 3 -> {
+                                            case "3" -> {
                                                 doneeList.replace(doneeToUpdate, doneeList.changeToOrganization(doneeToUpdate));
                                                 System.out.println(Green + "Category updated to organization successfully." + Reset);
                                             }
@@ -148,7 +146,7 @@ public class Test {
                                                 System.out.println(Red + "Invalid category. Please only enter '1', '2', or '3'." + Reset);
                                             }
                                         }
-                                    } while (newCategory < 1 || newCategory > 3);
+                                    } while (!"1".equals(choice) && !"2".equals(choice) && !"3".equals(choice));
                                 }
                                 default ->
                                     System.out.println(Red + "Invalid option. No changes made." + Reset);
@@ -157,103 +155,101 @@ public class Test {
                             System.out.println(Red + "Invalid ID. No donee found with the given ID." + Reset);
                         }
                     }
-                    case 4 -> {
-                        System.out.println("Search Donor by:");
-                        System.out.println("1. Donee ID");
-                        System.out.println("2. Donee Name \n");
-                        System.out.print("Enter your choice: ");
+                    case "4" -> {
+                        do {
+                            System.out.println("Search Donor by:");
+                            System.out.println("1. Donee ID");
+                            System.out.println("2. Donee Name \n");
+                            System.out.print("Enter your choice: ");
 
-                        String searchChoice = scanner.nextLine().trim();
+                            choice = scanner.nextLine().trim();
 
-                        switch (searchChoice) {
-                            case "1" -> {
-                                System.out.print("Enter donee ID to search: ");
-                                String id = scanner.nextLine().trim();
+                            switch (choice) {
+                                case "1" -> {
+                                    System.out.print("Enter donee ID to search: ");
+                                    String id = scanner.nextLine().trim();
 
-                                Donee donee = doneeList.findById(id);
-                                if (donee != null) {
-                                    System.out.println("\nFound donee: \n\n" + donee.toString());
-                                } else {
-                                    System.out.println("No donee found with ID: " + Red + id + Reset);
-                                }
-                            }
-                            case "2" -> {
-                                System.out.print("Enter donee name to search: ");
-                                String name = scanner.nextLine().trim().toLowerCase();
-
-                                LinkedList<Donee> matchingDonees = new LinkedList<>();
-                                Node<Donee> current = doneeList.head;
-
-                                while (current != null) {
-                                    if (current.data.getName().toLowerCase().contains(name)) {
-                                        matchingDonees.insert(current.data);
+                                    Donee donee = doneeList.findById(id);
+                                    if (donee != null) {
+                                        System.out.println(Green + "\nFound donee: " + Reset + donee.toString());
+                                    } else {
+                                        System.out.println(Red + "\nNo donee found with ID: " + id + Reset);
                                     }
-                                    current = current.next;
                                 }
+                                case "2" -> {
+                                    System.out.print("Enter donee name to search: ");
+                                    String name = scanner.nextLine().trim().toLowerCase();
 
-                                if (!matchingDonees.isEmpty()) {
-                                    System.out.println("\nMatching donees found: \n");
-                                    matchingDonees.show();
-                                } else {
-                                    System.out.println("No donees found with name containing: " + Red + name + Reset);
+                                    LinkedList<Donee> matchingDonees = new LinkedList<>();
+                                    Node<Donee> current = doneeList.head;
+
+                                    while (current != null) {
+                                        if (current.data.getName().toLowerCase().contains(name)) {
+                                            matchingDonees.insert(current.data);
+                                        }
+                                        current = current.next;
+                                    }
+
+                                    if (!matchingDonees.isEmpty()) {
+                                        System.out.print(Green + "\nMatching donees found: " + Reset);
+                                        matchingDonees.show();
+                                    } else {
+                                        System.out.println(Red + "\nNo donees found with name containing: " + name + Reset);
+                                    }
                                 }
+                                default ->
+                                    System.out.println(Red + "Invalid choice. Please enter '1' for ID search or '2' for Name search.\n " + Reset);
                             }
-                            default ->
-                                System.out.println(Red + "Invalid choice. Please enter '1' for ID search or '2' for Name search." + Reset);
-                        }
+                        } while (!"1".equals(choice) && !"2".equals(choice));
                     }
-
-                    case 5 -> {
+                    case "5" -> {
                         // List donees with all donations made
+                        System.out.print(Green + "List of all Donees: " + Reset);
                         doneeList.show();
                     }
-                    case 6 -> {
+                    case "6" -> {
                         // Filter donee based on criteria
-                        int filterChoice = 0;
                         do {
                             System.out.println("1. Filter by Individual");
                             System.out.println("2. Filter by Family");
                             System.out.println("3. Filter by Organization \n");
                             System.out.print("Enter your choice: ");
-                            filterChoice = scanner.nextInt();
-                            scanner.nextLine(); // consume the newline
-                            System.out.println("");
+                            choice = scanner.nextLine();
 
-                            switch (filterChoice) {
-                                case 1 -> {
+                            switch (choice) {
+                                case "1" -> {
                                     LinkedList<Individual> individualList = doneeList.filterByCategory(Individual.class);
+                                    System.out.print(Green + "\nList of all Individual Donees: " + Reset);
                                     individualList.show();
                                 }
-                                case 2 -> {
+                                case "2" -> {
                                     LinkedList<Family> familyList = doneeList.filterByCategory(Family.class);
+                                    System.out.print(Green + "\nList of all Family Donees: " + Reset);
                                     familyList.show();
                                 }
-                                case 3 -> {
+                                case "3" -> {
                                     LinkedList<Organization> organizationList = doneeList.filterByCategory(Organization.class);
+                                    System.out.print(Green + "\nList of all Organization Donees: " + Reset);
                                     organizationList.show();
                                 }
                                 default -> {
-                                    System.out.println(Red + "Invalid category. Please only enter '1', '2', or '3'." + Reset);
+                                    System.out.println(Red + "Invalid category. Please only enter '1', '2', or '3'.\n" + Reset);
                                 }
                             }
-                        } while (filterChoice < 1 || filterChoice > 3);
+                        } while (!"1".equals(choice) && !"2".equals(choice) && !"3".equals(choice));
                     }
-//                    case 7 -> {
-//                        // Generate summary reports
-//                        doneeList.generateSummaryReport();
-//                    }
-                    case 8 -> {
+                    case "7" -> {
                         // Exit the program
                         doneeList.saveToFile("donees.txt");
-                        System.out.println("Changes saved");
+                        System.out.println(Green + "Changes saved sucessfully..."+ Reset);
                     }
-                    case 9 -> {
+                    case "8" -> {
                         // Exit the program
                         running = false;
-                        System.out.println("Exiting...");
+                        System.out.println(Green + "Exiting..." + Reset);
                     }
                     default ->
-                        System.out.println("Invalid option, please try again.");
+                        System.out.println(Red + "Invalid option, please try again." + Reset);
                 }
 
                 System.out.println();
@@ -278,9 +274,5 @@ public class Test {
         }
 
         return prefix + String.format("%05d", maxId + 1);
-    }
-
-    private static boolean isValidName(String name) {
-        return name != null && name.matches("^[a-zA-Z\\s]+$") && name.length() >= 2 && name.length() <= 30;
     }
 }
